man: add ArchLinux(7)
This outlines what files needs to be configured to get a well-behaved
system.

For now it is extremely terse, but the main point is to work as a
check-list, and to point people to the correct manpages to learn more.

I'll be happy to take patches to extend this in the future.

One might argue that this sholud not be part of initscripts, and we
might indeed move it to 'filesystem' or something like that in the
future. For the time being I'll keep it here for convenience.

Suggested-by: Thomas Bächler <thomas@archlinux.org>
Signed-off-by: Tom Gundersen <teg@jklm.no>