arch-modules-load: drop from systemd compat layer
We now no longer support reading the MODULES array when booting
with systemd. The support is kept in initscripts (though a warning
is printed) for sake of compatibility.

Both initscripts and systemd should be using the new modules-load.d
syntax and we should remove all mention of the old from varios
documentation, for the sake of simplicity.

Signed-off-by: Tom Gundersen <teg@jklm.no>